split app into packages
To make the code easier to maintain, split it into multiple
packages and modules within those packages.

The top level package is "musicbingo". Inside that there are
the following packages:

* docgen - modules to create printable documents
* mp3    - modules to parse, edit and create MP3 files
* tests  - various unit tests
