+++ .pylintrc
@@ -485,10 +485,10 @@ max-attributes=7
 max-bool-expr=5
 
 # Maximum number of branch for function / method body
-max-branches=12
+max-branches=14
 
 # Maximum number of locals for function / method body
-max-locals=17
+max-locals=20
 
 # Maximum number of parents for a class (see R0901).
 max-parents=7
