+++ musicbingo/__main__.py
@@ -1,0 +1,11 @@
+"""
+Main entry point to the Music Bingo app.
+
+Start using
+
+pthon3 -m musicbingo
+"""
+
+from musicbingo.gui import main
+
+main()
