+++ musicbingo/mp3/__init__.py
@@ -1,0 +1,8 @@
+"""
+package containing the MP3 parsing and editing code
+"""
+
+from musicbingo.mp3.editor import MP3Editor
+from musicbingo.mp3.exceptions import InvalidMP3Exception, NotAnMP3Exception
+from musicbingo.mp3.factory import MP3Factory
+from musicbingo.mp3.parser import MP3Parser
