+++ tox.ini
@@ -12,5 +12,5 @@ deps =
 install_command = pip3 install {opts} {packages}
 
 commands =
-         pylint --rcfile=.pylintrc MusicBingo.py
-         mypy MusicBingo.py
+         pylint --rcfile=.pylintrc -j 0 musicbingo
+         mypy musicbingo
